pyproject.toml
[tool.black]
exclude = '''
/(
  | migrations
)/
'''

[tool.poetry]
name = "weni-ai"
version = "2.1.29"
description = "weni-ai: weni-ai-engine"
authors = ["weni.ai"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "~=3.8.13"
django = "~=3.2.6"
djangorestframework = "~=3.12.4"
django-filter = "~=2.4.0"
django-cors-headers = "~=3.7.0"
requests = "~=2.23.0"
coreapi = "~=2.3.3"
whitenoise = "~=5.3.0"
pytz = "~=2020.1"
drf-yasg2 = "~=1.19.4"
gunicorn = "~=19.9.0"
gevent = "~=1.4.0"
packaging = "~=20.4"
django-environ-2 = "~=2.1.0"
boto3 = "~=1.18.13"
sentry-sdk = "~=1.3.1"
redis = "*"
celery = "~=5.1.2"
django-celery-beat = "~=2.2.1"
django-celery-results = "~=2.2.0"
python-slugify = "~=4.0.0"
openpyxl = "~=3.0.3"
kombu = "~=5.1.0"
matplot = "~=0.1.9"
django-redis = "~=4.12.1"
elastic-apm = "~=6.3.3"
mozilla-django-oidc = "~=2.0.0"
djangorestframework-recaptcha = "~=0.2.0"
djangogrpcframework = "~=0.2.1"
grpcio = "~=1.39.0"
grpcio-tools = "~=1.39.0"
django-elasticsearch-dsl = "~=7.2.0"
django-elasticsearch-dsl-drf = "~=0.22.1"
elasticsearch = "<7.14.0"
elasticsearch-dsl = "~=7.4.0"
psycopg2-binary = "~=2.9.1"
weni-protobuffers = "~=1.2.1"
black = "21.7b0"
Pillow = "~=8.4.0"
django-csp = "^3.7"

[tool.poetry.dev-dependencies]
flake8 = "~=4.0.0"
requests-mock = "~=1.8.0"
coverage = "~=5.5"
ipython = "*"
black = "==21.7b0"
autopep8 = "^1.6.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
